requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
tensorflow
Pillow
requests
beautifulsoup4
tqdm
pymysql 
python-dotenv
fastapi
uvicorn
sqlalchemy
pymysql
python-multipart
opencv-python
pytesseract